feat: Add equipment editing feature
Added the ability to edit equipment names in the Material List component. Users can now click on the equipment name to transform it into an input field, make changes, and save the updates to the server. The "updatedEquipment" property was added to the Material model to track changes. When an equipment name is edited and saved, it is updated in the UI and sent to the server for persistence.